refactor(cli): Reduce file noise by moving tests from resolver_v2.rs into dedicated test file
